refactor: Review the DecoratesInput implementation
- Move the compatibility layer to a dedicated internal namespace
- Add a few more comments and fix unnecessary PHPDoc
