+++ src/Input/Compatibility/DecoratesInputSymfony5.php
@@ -21,13 +21,14 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace Fidry\Console\Input;
+namespace Fidry\Console\Input\Compatibility;
 
 use Symfony\Component\Console\Input\InputDefinition;
 use Symfony\Component\Console\Input\InputInterface;
 use function func_get_args;
 
 /**
+ * @internal
  * @psalm-require-implements InputInterface
  */
 trait DecoratesInputSymfony5
@@ -97,9 +98,6 @@ trait DecoratesInputSymfony5
         return $this->input->getOption(...func_get_args());
     }
 
-    /**
-     * @param non-empty-string $name
-     */
     public function hasOption(string $name, bool $onlyRealParams = false): bool
     {
         return $this->input->hasOption(...func_get_args());
